Eliminate runtime references into the stdlib C primitives
Move the runtime-service code that still lived in the stdlib C
primitive files back into runtime/, so that the two halves of the C
sources are cleanly layered: library code calls into the runtime,
never the reverse.  Verified by checking that no member of the
runtime libraries has an undefined symbol defined in a stdlib/prims
member, across all bytecode/native and normal/debug/instrumented
variants.

- str.c: caml_string_length, caml_string_is_c_safe and
  caml_alloc_sprintf return to runtime/str.c (used by intern, extern,
  startup, sys and runtime_events).
- floats.c: the float-formatting locale (caml_locale,
  caml_init_locale, caml_free_locale) and caml_log1p return to
  runtime/floats.c (used by startup and memprof).
- ints.c, float32.c, bigarray.c: the custom operations and their
  supporting machinery (including bigarray allocation and layout)
  return to runtime/{ints,float32,bigarray}.c; custom.c registers
  these operations for the marshaller.  The float16 conversions,
  needed by both halves of bigarray, move to the runtime-internal
  header runtime/float16.h.
- md5.c: the MD5 core returns to runtime/md5.c (codefrag digests).
- array.c: caml_array_length and caml_is_double_array return to
  runtime/array.c (used by bytecode backtraces).
- hash.c: the caml_hash_mix_* functions return to runtime/hash.c
  (used by the custom operations); the MurmurHash mixing macros move
  to an internals section of caml/hash.h, shared with the generic
  hashing primitives.
- The object-id primitives return to runtime/obj.c (intern and
  dynamic.c allocate ids).

A stdlib/prims file whose runtime counterpart keeps the original name
is renamed with a _prims suffix, since archive members need distinct
basenames.

The runtime keeps one deliberate set of references into stdlib/prims:
a force-link table in startup_nat.c referencing one primitive per
file.  Previously, runtime-internal references such as custom.c ->
caml_float32_ops dragged whole primitive objects into every native
executable, so dynamically-loaded OCaml code could rely on the stdlib
primitives being present even when the main program does not use
them.  The table makes that guarantee explicit, mirroring the
bytecode runtime's builtin primitive table (prims.c).
