+++ stdlib/prims/bigarray_prims.c
@@ -35,620 +35,7 @@
 #define int16 caml_ba_int16
 #define uint16 caml_ba_uint16
 
-/* Half-precision floating point numbers */
-
-#if defined(__GNUC__) && defined(__aarch64__)
-
-union float16_bits { uint16_t i; _Float16 f; };
-
-Caml_inline float caml_float16_to_float(uint16 d)
-{
-  union float16_bits u;
-  u.i = d; return u.f;
-}
-
-Caml_inline uint16 caml_float_to_float16(float d)
-{
-  union float16_bits u;
-  u.f = d; return u.i;
-}
-
-#elif defined(__GNUC__) && defined(__F16C__)
-
-#include <immintrin.h>
-
-Caml_inline float caml_float16_to_float(uint16 d)
-{ return _cvtsh_ss(d); }
-
-Caml_inline uint16 caml_float_to_float16(float d)
-{ return _cvtss_sh(d, (_MM_FROUND_TO_NEAREST_INT | _MM_FROUND_NO_EXC)); }
-
-#else
-
-union float_bits {
-  uint32_t i;
-  float f;
-};
-
-/*
- * half_to_float_fast5
- * https://gist.github.com/rygorous/2144712
- */
-static float caml_float16_to_float(uint16 d)
-{
-  static const union float_bits magic = { (254 - 15) << 23 };
-  static const union float_bits was_infnan = { (127 + 16) << 23 };
-
-  union float_bits o;
-
-  o.i = (d & 0x7fff) << 13;     /* exponent/mantissa bits */
-  o.f *= magic.f;               /* exponent adjust */
-  if (o.f >= was_infnan.f)      /* make sure Inf/NaN survive */
-    o.i |= 255 << 23;
-  o.i |= (d & 0x8000) << 16;    /* sign bit */
-  return o.f;
-}
-
-/*
- * float_to_half_fast3_rtne
- * https://gist.github.com/rygorous/2156668
- */
-static uint16 caml_float_to_float16(float d)
-{
-  static const union float_bits f32infty = { 255 << 23 };
-  static const union float_bits f16max = { (127 + 16) << 23 };
-  static const union float_bits denorm_magic =
-    { ((127 - 15) + (23 - 10) + 1) << 23 };
-  static const uint32_t sign_mask = 0x80000000u;
-
-  union float_bits f;
-  uint16 o = 0;
-  uint32_t sign;
-
-  f.f = d;
-  sign = f.i & sign_mask;
-  f.i ^= sign;
-
-  // NOTE all the integer compares in this function can be safely
-  // compiled into signed compares since all operands are below
-  // 0x80000000. Important if you want fast straight SSE2 code
-  // (since there's no unsigned PCMPGTD).
-
-  if (f.i >= f16max.i) // result is Inf or NaN (all exponent bits set)
-    o = (f.i > f32infty.i) ? 0x7e00 : 0x7c00; // NaN->qNaN and Inf->Inf
-  else // (De)normalized number or zero
-  {
-    if (f.i < (113 << 23)) // resulting FP16 is subnormal or zero
-    {
-      // use a magic value to align our 10 mantissa bits at the bottom of
-      // the float. as long as FP addition is round-to-nearest-even this
-      // just works.
-      f.f += denorm_magic.f;
-
-      // and one integer subtract of the bias later, we have our final float!
-      o = f.i - denorm_magic.i;
-    }
-    else
-    {
-      uint32_t mant_odd = (f.i >> 13) & 1; // resulting mantissa is odd
-
-      // update exponent, rounding bias part 1
-      f.i += ((uint32_t)(15 - 127) << 23) + 0xfff;
-      // rounding bias part 2
-      f.i += mant_odd;
-      // take the bits!
-      o = f.i >> 13;
-    }
-  }
-
-  o |= sign >> 16;
-  return o;
-}
-
-#endif  /* defined(__GNUC__) && defined(__F16C__) */
-
-CAMLexport double caml_double_of_float16(intnat x)
-{
-  return (double) caml_float16_to_float((uint16) x);
-}
-
-CAMLexport intnat caml_float16_of_double(double x)
-{
-  return (intnat) caml_float_to_float16((float) x);
-}
-
-Caml_inline uint32_t caml_hash_mix_float16(uint32_t hash, uint16 d)
-{
-  /* Normalize NaNs */
-  if ((d & 0x7c00) == 0x7c00 && (d & 0x03ff) != 0) {
-    d = 0x7c01;
-  }
-  /* Normalize -0 into +0 */
-  else if (d == 0x8000) {
-    d = 0;
-  }
-  return caml_hash_mix_uint32(hash, d);
-}
-
-/* Compute the number of elements of a big array */
-
-CAMLexport uintnat caml_ba_num_elts(struct caml_ba_array * b)
-{
-  uintnat num_elts;
-  num_elts = 1;
-  for (int i = 0; i < b->num_dims; i++) num_elts = num_elts * b->dim[i];
-  return num_elts;
-}
-
-/* Size in bytes of a bigarray element, indexed by bigarray kind */
-
-CAMLexport const int caml_ba_element_size[] =
-{ 4 /*FLOAT32*/, 8 /*FLOAT64*/,
-  1 /*SINT8*/, 1 /*UINT8*/,
-  2 /*SINT16*/, 2 /*UINT16*/,
-  4 /*INT32*/, 8 /*INT64*/,
-  sizeof(value) /*CAML_INT*/, sizeof(value) /*NATIVE_INT*/,
-  8 /*COMPLEX32*/, 16 /*COMPLEX64*/,
-  1 /*CHAR*/, 2 /*FLOAT16*/
-};
-
-/* Compute the number of bytes for the elements of a big array */
-
-CAMLexport uintnat caml_ba_byte_size(struct caml_ba_array * b)
-{
-  return caml_ba_num_elts(b)
-         * caml_ba_element_size[b->flags & CAML_BA_KIND_MASK];
-}
-
-/* Operation table for bigarrays */
-
-CAMLexport const struct custom_operations caml_ba_ops = {
-  "_bigarr02",
-  caml_ba_finalize,
-  caml_ba_compare,
-  caml_ba_hash,
-  caml_ba_serialize,
-  caml_ba_deserialize,
-  custom_compare_ext_default,
-  custom_fixed_length_default
-};
-
-/* Allocation of a big array */
-
-/* [caml_ba_alloc] will allocate a new bigarray object in the heap.
-   If [data] is NULL, the memory for the contents is also allocated
-   (with [malloc]) by [caml_ba_alloc].
-   [data] cannot point into the OCaml heap.
-   [dim] may point into an object in the OCaml heap.
-*/
-CAMLexport value
-caml_ba_alloc(int flags, int num_dims, void * data, intnat * dim)
-{
-  uintnat num_elts, asize, size;
-  int uses_resources;
-  value res;
-  struct caml_ba_array * b;
-  intnat dimcopy[CAML_BA_MAX_NUM_DIMS];
-
-  CAMLassert(0 <= num_dims);
-  CAMLassert(num_dims <= CAML_BA_MAX_NUM_DIMS);
-  CAMLassert((flags & CAML_BA_KIND_MASK) < CAML_BA_FIRST_UNIMPLEMENTED_KIND);
-  for (int i = 0; i < num_dims; i++) dimcopy[i] = dim[i];
-  num_elts = 1;
-  for (int i = 0; i < num_dims; i++) {
-    if (caml_umul_overflow(num_elts, dimcopy[i], &num_elts))
-      caml_raise_out_of_memory();
-  }
-  if (caml_umul_overflow(num_elts,
-                         caml_ba_element_size[flags & CAML_BA_KIND_MASK],
-                         &size))
-    caml_raise_out_of_memory();
-  if (data == NULL) {
-    data = malloc(size);
-    if (data == NULL && size != 0) caml_raise_out_of_memory();
-    flags |= CAML_BA_MANAGED;
-  }
-  asize = SIZEOF_BA_ARRAY + num_dims * sizeof(intnat);
-  uses_resources =
-    ((flags & CAML_BA_MANAGED_MASK) == CAML_BA_MANAGED)
-    && !(flags & CAML_BA_SUBARRAY);
-  res = caml_alloc_custom_dep(&caml_ba_ops, asize, uses_resources ? size : 0);
-  b = Caml_ba_array_val(res);
-  b->data = data;
-  b->num_dims = num_dims;
-  b->flags = flags;
-  b->proxy = NULL;
-  for (int i = 0; i < num_dims; i++) b->dim[i] = dimcopy[i];
-  return res;
-}
-
-/* Same as caml_ba_alloc, but dimensions are passed as a list of
-   arguments */
-
-CAMLexport value caml_ba_alloc_dims(int flags, int num_dims, void * data, ...)
-{
-  va_list ap;
-  intnat dim[CAML_BA_MAX_NUM_DIMS];
-  value res;
-
-  CAMLassert(num_dims <= CAML_BA_MAX_NUM_DIMS);
-  va_start(ap, data);
-  for (int i = 0; i < num_dims; i++) dim[i] = va_arg(ap, intnat);
-  va_end(ap);
-  res = caml_ba_alloc(flags, num_dims, data, dim);
-  return res;
-}
-
-/* Finalization of a big array */
-
-CAMLexport void caml_ba_finalize(value v)
-{
-  struct caml_ba_array * b = Caml_ba_array_val(v);
-
-  switch ((enum caml_ba_managed)(b->flags & CAML_BA_MANAGED_MASK)) {
-  case CAML_BA_EXTERNAL:
-    break;
-  case CAML_BA_MANAGED:
-    if (b->proxy == NULL) {
-      free(b->data);
-      caml_free_dependent_memory(v, caml_ba_byte_size(b));
-    } else {
-      if (caml_atomic_counter_decr(&b->proxy->refcount) == 0) {
-        free(b->proxy->data);
-        caml_free_dependent_memory(v, b->proxy->size);
-        free(b->proxy);
-      }
-    }
-    break;
-  case CAML_BA_MAPPED_FILE:
-    /* Bigarrays for mapped files use a different finalization method */
-    CAMLunreachable();
-  }
-}
-
-/* Comparison of two big arrays */
-
-CAMLexport int caml_ba_compare(value v1, value v2)
-{
-  struct caml_ba_array * b1 = Caml_ba_array_val(v1);
-  struct caml_ba_array * b2 = Caml_ba_array_val(v2);
-  uintnat num_elts;
-  intnat flags1, flags2;
-
-  /* Compare kind & layout in case the arguments are of different types */
-  flags1 = b1->flags & (CAML_BA_KIND_MASK | CAML_BA_LAYOUT_MASK);
-  flags2 = b2->flags & (CAML_BA_KIND_MASK | CAML_BA_LAYOUT_MASK);
-  if (flags1 != flags2) return flags2 - flags1;
-  /* Compare number of dimensions */
-  if (b1->num_dims != b2->num_dims) return b2->num_dims - b1->num_dims;
-  /* Same number of dimensions: compare dimensions lexicographically */
-  for (int i = 0; i < b1->num_dims; i++) {
-    intnat d1 = b1->dim[i];
-    intnat d2 = b2->dim[i];
-    if (d1 != d2) return d1 < d2 ? -1 : 1;
-  }
-  /* Same dimensions: compare contents lexicographically */
-  num_elts = caml_ba_num_elts(b1);
-
-#define DO_INTEGER_COMPARISON(type) \
-  { const type * p1 = b1->data; const type * p2 = b2->data; \
-    for (uintnat n = 0; n < num_elts; n++) { \
-      type e1 = *p1++; type e2 = *p2++; \
-      if (e1 < e2) return -1; \
-      if (e1 > e2) return 1; \
-    } \
-    return 0; \
-  }
-#define DO_GENERIC_UNORDERED_COMPARISON(ptype, etype, conv) \
-  { const ptype * p1 = b1->data; const ptype * p2 = b2->data; \
-    for (uintnat n = 0; n < num_elts; n++) { \
-      etype e1 = conv(*p1++); etype e2 = conv(*p2++); \
-      if (e1 < e2) return -1; \
-      if (e1 > e2) return 1; \
-      if (e1 != e2) { \
-        Caml_state->compare_unordered = 1; \
-        if (e1 == e1) return 1; \
-        if (e2 == e2) return -1; \
-      } \
-    } \
-    return 0; \
-  }
-#define DO_FLOAT_COMPARISON(type) \
-  DO_GENERIC_UNORDERED_COMPARISON(type, type, )
-
-  switch ((enum caml_ba_kind)(b1->flags & CAML_BA_KIND_MASK)) {
-  case CAML_BA_FLOAT16:
-    DO_GENERIC_UNORDERED_COMPARISON(uint16, float, caml_float16_to_float);
-  case CAML_BA_COMPLEX32:
-    num_elts *= 2; fallthrough;
-  case CAML_BA_FLOAT32:
-    DO_FLOAT_COMPARISON(float);
-  case CAML_BA_COMPLEX64:
-    num_elts *= 2; fallthrough;
-  case CAML_BA_FLOAT64:
-    DO_FLOAT_COMPARISON(double);
-  case CAML_BA_CHAR:
-    DO_INTEGER_COMPARISON(caml_ba_uint8);
-  case CAML_BA_SINT8:
-    DO_INTEGER_COMPARISON(caml_ba_int8);
-  case CAML_BA_UINT8:
-    DO_INTEGER_COMPARISON(caml_ba_uint8);
-  case CAML_BA_SINT16:
-    DO_INTEGER_COMPARISON(caml_ba_int16);
-  case CAML_BA_UINT16:
-    DO_INTEGER_COMPARISON(caml_ba_uint16);
-  case CAML_BA_INT32:
-    DO_INTEGER_COMPARISON(int32_t);
-  case CAML_BA_INT64:
-    DO_INTEGER_COMPARISON(int64_t);
-  case CAML_BA_CAML_INT:
-  case CAML_BA_NATIVE_INT:
-    DO_INTEGER_COMPARISON(intnat);
-  default: CAMLunreachable();
-  }
-#undef DO_INTEGER_COMPARISON
-#undef DO_FLOAT_COMPARISON
-}
-
-/* Hashing of a bigarray */
-
-CAMLexport intnat caml_ba_hash(value v)
-{
-  struct caml_ba_array * b = Caml_ba_array_val(v);
-  intnat num_elts;
-  uint32_t h, w;
-
-  num_elts = 1;
-  for (int i = 0; i < b->num_dims; i++) num_elts = num_elts * b->dim[i];
-  h = 0;
-
-  switch ((enum caml_ba_kind)(b->flags & CAML_BA_KIND_MASK)) {
-  case CAML_BA_CHAR:
-  case CAML_BA_SINT8:
-  case CAML_BA_UINT8: {
-    caml_ba_uint8 * p = b->data;
-    if (num_elts > 256) num_elts = 256;
-    for (intnat n = 0; n + 4 <= num_elts; n += 4, p += 4) {
-      w = p[0] | (p[1] << 8) | (p[2] << 16) | (p[3] << 24);
-      h = caml_hash_mix_uint32(h, w);
-    }
-    w = 0;
-    switch (num_elts & 3) {
-    case 3: w  = p[2] << 16; fallthrough;
-    case 2: w |= p[1] << 8;  fallthrough;
-    case 1: w |= p[0];
-            h = caml_hash_mix_uint32(h, w);
-    }
-    break;
-  }
-  case CAML_BA_SINT16:
-  case CAML_BA_UINT16: {
-    caml_ba_uint16 * p = b->data;
-    if (num_elts > 128) num_elts = 128;
-    for (intnat n = 0; n + 2 <= num_elts; n += 2, p += 2) {
-      w = p[0] | (p[1] << 16);
-      h = caml_hash_mix_uint32(h, w);
-    }
-    if ((num_elts & 1) != 0)
-      h = caml_hash_mix_uint32(h, p[0]);
-    break;
-  }
-  case CAML_BA_INT32:
-  {
-    uint32_t * p = b->data;
-    if (num_elts > 64) num_elts = 64;
-    for (intnat n = 0; n < num_elts; n++, p++) h = caml_hash_mix_uint32(h, *p);
-    break;
-  }
-  case CAML_BA_CAML_INT:
-  case CAML_BA_NATIVE_INT:
-  {
-    intnat * p = b->data;
-    if (num_elts > 64) num_elts = 64;
-    for (intnat n = 0; n < num_elts; n++, p++) h = caml_hash_mix_intnat(h, *p);
-    break;
-  }
-  case CAML_BA_INT64:
-  {
-    int64_t * p = b->data;
-    if (num_elts > 32) num_elts = 32;
-    for (intnat n = 0; n < num_elts; n++, p++) h = caml_hash_mix_int64(h, *p);
-    break;
-  }
-  case CAML_BA_FLOAT16:
-  {
-    uint16 * p = b->data;
-    if (num_elts > 128) num_elts = 128;
-    for (intnat n = 0; n < num_elts; n++, p++) h = caml_hash_mix_float16(h, *p);
-    break;
-  }
-  case CAML_BA_COMPLEX32:
-    num_elts *= 2;
-    fallthrough;
-  case CAML_BA_FLOAT32:
-  {
-    float * p = b->data;
-    if (num_elts > 64) num_elts = 64;
-    for (intnat n = 0; n < num_elts; n++, p++) h = caml_hash_mix_float(h, *p);
-    break;
-  }
-  case CAML_BA_COMPLEX64:
-    num_elts *= 2;
-    fallthrough;
-  case CAML_BA_FLOAT64:
-  {
-    double * p = b->data;
-    if (num_elts > 32) num_elts = 32;
-    for (intnat n = 0; n < num_elts; n++, p++) h = caml_hash_mix_double(h, *p);
-    break;
-  }
-  default: CAMLunreachable();
-  }
-  return h;
-}
-
-static void caml_ba_serialize_longarray(void * data,
-                                        intnat num_elts,
-                                        intnat min_val, intnat max_val)
-{
-#ifdef ARCH_SIXTYFOUR
-  int overflow_32 = 0;
-  for (intnat n = 0, *p = data; n < num_elts; n++, p++) {
-    if (*p < min_val || *p > max_val) { overflow_32 = 1; break; }
-  }
-  if (overflow_32) {
-    caml_serialize_int_1(1);
-    caml_serialize_block_8(data, num_elts);
-  } else {
-    caml_serialize_int_1(0);
-    for (intnat n = 0, *p = data; n < num_elts; n++, p++)
-      caml_serialize_int_4((int32_t) *p);
-  }
-#else
-  caml_serialize_int_1(0);
-  caml_serialize_block_4(data, num_elts);
-#endif
-}
-
-CAMLexport void caml_ba_serialize(value v,
-                              uintnat * wsize_32,
-                              uintnat * wsize_64)
-{
-  struct caml_ba_array * b = Caml_ba_array_val(v);
-  intnat num_elts;
-
-  /* Serialize header information */
-  caml_serialize_int_4(b->num_dims);
-  caml_serialize_int_4(b->flags & (CAML_BA_KIND_MASK | CAML_BA_LAYOUT_MASK));
-  for (int i = 0; i < b->num_dims; i++) {
-    intnat len = b->dim[i];
-    if (len < 0xffff) {
-      caml_serialize_int_2(len);
-    } else {
-      caml_serialize_int_2(0xffff);
-      caml_serialize_int_8(len);
-    }
-  }
-  /* Compute total number of elements */
-  num_elts = 1;
-  for (int i = 0; i < b->num_dims; i++) num_elts = num_elts * b->dim[i];
-  /* Serialize elements */
-  switch ((enum caml_ba_kind)(b->flags & CAML_BA_KIND_MASK)) {
-  case CAML_BA_CHAR:
-  case CAML_BA_SINT8:
-  case CAML_BA_UINT8:
-    caml_serialize_block_1(b->data, num_elts); break;
-  case CAML_BA_SINT16:
-  case CAML_BA_UINT16:
-  case CAML_BA_FLOAT16:
-    caml_serialize_block_2(b->data, num_elts); break;
-  case CAML_BA_FLOAT32:
-  case CAML_BA_INT32:
-    caml_serialize_block_4(b->data, num_elts); break;
-  case CAML_BA_COMPLEX32:
-    caml_serialize_block_4(b->data, num_elts * 2); break;
-  case CAML_BA_FLOAT64:
-  case CAML_BA_INT64:
-    caml_serialize_block_8(b->data, num_elts); break;
-  case CAML_BA_COMPLEX64:
-    caml_serialize_block_8(b->data, num_elts * 2); break;
-  case CAML_BA_CAML_INT:
-    caml_ba_serialize_longarray(b->data, num_elts, INT32_MIN/2, INT32_MAX/2);
-    break;
-  case CAML_BA_NATIVE_INT:
-    caml_ba_serialize_longarray(b->data, num_elts, INT32_MIN, INT32_MAX);
-    break;
-  default: CAMLunreachable();
-  }
-  /* Compute required size in OCaml heap.  Assumes struct caml_ba_array
-     is exactly 4 + num_dims words */
-  static_assert(SIZEOF_BA_ARRAY == 4 * sizeof(value), "");
-  *wsize_32 = (4 + b->num_dims) * 4;
-  *wsize_64 = (4 + b->num_dims) * 8;
-}
-
-static void caml_ba_deserialize_longarray(void * dest, intnat num_elts)
-{
-  int sixty = caml_deserialize_uint_1();
-#ifdef ARCH_SIXTYFOUR
-  if (sixty) {
-    caml_deserialize_block_8(dest, num_elts);
-  } else {
-    for (intnat n = 0, *p = dest; n < num_elts; n++, p++)
-      *p = caml_deserialize_sint_4();
-  }
-#else
-  if (sixty)
-    caml_deserialize_error("input_value: cannot read bigarray "
-                      "with 64-bit OCaml ints");
-  caml_deserialize_block_4(dest, num_elts);
-#endif
-}
-
-CAMLexport uintnat caml_ba_deserialize(void * dst)
-{
-  struct caml_ba_array * b = dst;
-  uintnat num_elts, size;
-
-  /* Read back header information */
-  b->num_dims = caml_deserialize_uint_4();
-  if (b->num_dims < 0 || b->num_dims > CAML_BA_MAX_NUM_DIMS)
-    caml_deserialize_error("input_value: wrong number of bigarray dimensions");
-  b->flags = caml_deserialize_uint_4() | CAML_BA_MANAGED;
-  b->proxy = NULL;
-  for (int i = 0; i < b->num_dims; i++) {
-    intnat len = caml_deserialize_uint_2();
-    if (len == 0xffff) len = caml_deserialize_uint_8();
-    b->dim[i] = len;
-  }
-  /* Compute total number of elements.  Watch out for overflows (MPR#7765). */
-  num_elts = 1;
-  for (int i = 0; i < b->num_dims; i++) {
-    if (caml_umul_overflow(num_elts, b->dim[i], &num_elts))
-      caml_deserialize_error("input_value: size overflow for bigarray");
-  }
-  /* Determine array size in bytes.  Watch out for overflows (MPR#7765). */
-  if ((b->flags & CAML_BA_KIND_MASK) >= CAML_BA_FIRST_UNIMPLEMENTED_KIND)
-    caml_deserialize_error("input_value: bad bigarray kind");
-  if (caml_umul_overflow(num_elts,
-                         caml_ba_element_size[b->flags & CAML_BA_KIND_MASK],
-                         &size))
-    caml_deserialize_error("input_value: size overflow for bigarray");
-  /* Allocate room for data */
-  b->data = malloc(size);
-  caml_alloc_dependent_memory(Custom_val_data (dst), size);
-  if (b->data == NULL)
-    caml_deserialize_error("input_value: out of memory for bigarray");
-  /* Read data */
-  switch ((enum caml_ba_kind)(b->flags & CAML_BA_KIND_MASK)) {
-  case CAML_BA_CHAR:
-  case CAML_BA_SINT8:
-  case CAML_BA_UINT8:
-    caml_deserialize_block_1(b->data, num_elts); break;
-  case CAML_BA_SINT16:
-  case CAML_BA_UINT16:
-  case CAML_BA_FLOAT16:
-    caml_deserialize_block_2(b->data, num_elts); break;
-  case CAML_BA_FLOAT32:
-  case CAML_BA_INT32:
-    caml_deserialize_block_4(b->data, num_elts); break;
-  case CAML_BA_COMPLEX32:
-    caml_deserialize_block_4(b->data, num_elts * 2); break;
-  case CAML_BA_FLOAT64:
-  case CAML_BA_INT64:
-    caml_deserialize_block_8(b->data, num_elts); break;
-  case CAML_BA_COMPLEX64:
-    caml_deserialize_block_8(b->data, num_elts * 2); break;
-  case CAML_BA_CAML_INT:
-  case CAML_BA_NATIVE_INT:
-    caml_ba_deserialize_longarray(b->data, num_elts); break;
-  default: CAMLunreachable();
-  }
-  /* PR#5516: use C99's flexible array types if possible */
-  return SIZEOF_BA_ARRAY + b->num_dims * sizeof(intnat);
-}
-
-/* Allocate a bigarray from OCaml */
+#include "float16.h"
 
 CAMLprim value caml_ba_create(value vkind, value vlayout, value vdim)
 {
