+++ stdlib/prims/float32_prims.c
@@ -66,69 +66,6 @@ static_assert(sizeof(float) == sizeof(int32_t), "");
 #define Max_array_wosize                 (Max_wosize)
 #define Max_unboxed_float32_array_wosize (Max_array_wosize * (sizeof(intnat) / sizeof(float)))
 
-intnat caml_float32_compare_unboxed(float f, float g)
-{
-  /* If one or both of f and g is NaN, order according to the convention
-     NaN = NaN and NaN < x for all other floats x. */
-  /* This branchless implementation is from GPR#164.
-     Note that [f == f] if and only if f is not NaN.
-     We expand each subresult of the expression to
-     avoid sign-extension on 64bit. GPR#2250.  */
-  intnat res =
-      (intnat)(f > g) - (intnat)(f < g) + (intnat)(f == f) - (intnat)(g == g);
-  return res;
-}
-
-static int float32_cmp(value v1, value v2)
-{
-  return caml_float32_compare_unboxed(Float32_val(v1), Float32_val(v2));
-}
-
-static intnat float32_hash(value v)
-{
-  union {
-    float f;
-    uint32_t i;
-  } u;
-  uint32_t n;
-  u.f = Float32_val(v);  n = u.i;
-  /* Normalize NaNs */
-  if ((n & 0x7F800000) == 0x7F800000 && (n & 0x007FFFFF) != 0) {
-    n = 0x7F800001;
-  }
-  /* Normalize -0 into +0 */
-  else if (n == 0x80000000) {
-    n = 0;
-  }
-  return n;
-}
-
-static uintnat float32_deserialize(void *dst)
-{
-  *((float *)dst) = caml_deserialize_float_4();
-  return 4;
-}
-
-static void float32_serialize(value v, uintnat *bsize_32,
-                              uintnat *bsize_64)
-{
-  caml_serialize_float_4(Float32_val(v));
-  *bsize_32 = *bsize_64 = 4;
-}
-
-static const struct custom_fixed_length float32_length = {4, 4};
-
-CAMLexport const struct custom_operations caml_float32_ops = {
-  "_f32",
-  custom_finalize_default,
-  float32_cmp,
-  float32_hash,
-  float32_serialize,
-  float32_deserialize,
-  custom_compare_ext_default,
-  &float32_length
-};
-
 CAMLexport value caml_copy_float32(float f)
 {
   value res = caml_alloc_custom(&caml_float32_ops, 4, 0, 1);
@@ -190,6 +127,9 @@ CAMLprim value caml_sqrt_float32(value f)
 {
   return caml_copy_float32(sqrtf(Float32_val(f)));
 }
+
+extern intnat caml_float32_compare_unboxed(float, float);
+                                             /* see runtime/float32.c */
 
 CAMLprim value caml_float32_compare(value vf, value vg)
 {
